+++ tvm_client/src/debot/json_lib_utils.rs
@@ -12,6 +12,7 @@ use tvm_abi::token::Tokenizer;
 
 use crate::boc::internal::deserialize_cell_from_base64;
 use crate::boc::internal::serialize_cell_to_base64;
+
 #[derive(Serialize_repr, Deserialize_repr)]
 #[repr(u8)]
 #[derive(Default)]
@@ -63,8 +64,7 @@ impl Value {
     }
 
     fn new_string(v: String) -> Option<Self> {
-        let mut val = Self::default();
-        val.kind = ValKind::String;
+        let mut val = Self { kind: ValKind::String, ..Default::default() };
         if deserialize_cell_from_base64(&v, "QueryValue").is_ok() {
             val.value = v;
             val.kind = ValKind::Cell;
@@ -75,8 +75,7 @@ impl Value {
     }
 
     fn new_object(map: serde_json::map::Map<String, JsonValue>) -> Option<Self> {
-        let mut val = Self::default();
-        val.kind = ValKind::Object;
+        let mut val = Self { kind: ValKind::Object, ..Default::default() };
         for (k, v) in map {
             let json: JsonValue = serde_json::to_value(pack(v)?).ok()?;
             let packed = Self::pack_value_to_cell(json, Some(&k))?;
@@ -89,8 +88,7 @@ impl Value {
     }
 
     fn new_array(array: Vec<JsonValue>) -> Option<Self> {
-        let mut val = Self::default();
-        val.kind = ValKind::Array;
+        let mut val = Self { kind: ValKind::Array, ..Default::default() };
         for element in array {
             let json: JsonValue = serde_json::to_value(pack(element)?).ok()?;
             let packed = Self::pack_value_to_cell(json, None)?;